Limited data

+++ src/screens/HomeScreen/optionModal.js
@@ -1,51 +1,53 @@
-import { View, Text, StyleSheet, Image, TouchableOpacity } from 'react-native'
-import React, { useEffect, useState } from 'react';
-import {MaterialCommunityIcons} from '@expo/vector-icons';
+import { View, Text, StyleSheet, TouchableOpacity,Modal } from 'react-native'
+import React, { useState } from 'react';
+import {MaterialCommunityIcons,Feather,Entypo,MaterialIcons } from '@expo/vector-icons';
 import {useNavigation} from '@react-navigation/native';
-import { getLocalData } from '../../apis/GetLocalData';
 import { deletePost } from '../../../redux/reducers/postAction';
 import { useDispatch } from 'react-redux';
 import { coinTransfer } from '../../../redux/reducers/coinSlice';
-import { BlockUserApi, SavePostApi } from '../../../redux/reducers/ALL_APIs';
+import { BlockUserApi, followApi, SavePostApi } from '../../../redux/reducers/ALL_APIs';
 import {Ionicons} from '@expo/vector-icons';
-import { getSavedPostsApi } from '../../../redux/reducers/SettingsSlice';
 
+const OptionModal = ({modalVisible,item,deletePostID,BlockId,setModalVisible,resData}) => {
+  const [toggle, setToggle] = useState(false);
+  const [savedPost, setSavedPost] = useState(item?.saved_status);
+  const [follow, setFollow] = useState(item?.follow_status);
+  const [okay, setOkayCondition] = useState(null);
 
-const OptionModal = ({modalVisible,id,postId,deletePostID,BlockId,setModalVisible,saveStatus}) => {
-  const [userData, setUserData] = useState();
-  const [savedPost, setSavedPost] = useState(saveStatus);
   const dispatch    = useDispatch();
   const navigation  = useNavigation();
-  const getId = async () => {
-    getLocalData('USER_INFO').then( async (res) => {
-      const reData = res?.data;
-      const savedResult = await dispatch(getSavedPostsApi({user_id:res?.data?.id}))
-      // const saved = savedResult.payload.map(d => d.post_id == postId);
-      setUserData(reData);
-    });
-  }
-  
-  useEffect(() => {
-    getId();
-  },[])
 
   const handleDeletePost = async ()=>{
-    const postDetails = {post_id:postId}
-    const result      = await dispatch(deletePost(postDetails));
+    setOkayCondition("delete");
+    setModalVisible(false);
+    setToggle(true);
+  }
+
+  const handleOkay = async () =>{
+    if(okay == "delete"){
+      const postDetails = {post_id:item?.post_id}
+      const result      = await dispatch(deletePost(postDetails));
       if(result.payload.status  == 'Success'){
-        const coinDetails = {task:15, receiverId:0, senderId:id} 
+        const coinDetails = {task:15, receiverId:0, senderId:item?.id} 
         const coinResult  = await dispatch(coinTransfer(coinDetails));
         if(coinResult.payload.status  == 'Success'){
-          deletePostID(postId);
+          deletePostID(item?.post_id);
+          setToggle(false);
         }
       }
+   }else if(okay == "block"){
+    const postDetails = {fromuserid:resData?.id,touserid:item?.id};
+    const blockPostResult  = await dispatch(BlockUserApi(postDetails));
+    if(blockPostResult?.payload?.status == "Success"){
+      setToggle(false);
+      BlockId(item?.id);
     }
+   }
+  }
 
   const SavedPostHandle = async () => {
-    const postDetails = {user_id:userData?.id, post_id:postId};
-    console.log(postDetails);
+    const postDetails = {user_id:resData?.id, post_id:item?.post_id};
     const savedPostResult = await dispatch(SavePostApi(postDetails));
-    // console.log("savedPostResult",savedPostResult.payload);
     if(savedPostResult.payload.status == "Saved"){
       setSavedPost(true);
     }else{
@@ -55,65 +57,89 @@ const OptionModal = ({modalVisible,id,postId,deletePostID,BlockId,setModalVisibl
 
   const handleReport = () => {
     setModalVisible(false);
-    navigation.navigate('ReportPost', {
-      postId, id})
+    navigation.navigate('ReportPost', {postId: item?.post_id, id: item?.id})
   }
 
-
-  const handleUnfollow = () => {
-
-  }
-
-  const BlockPostHandle = async () => {
-    const postDetails = {fromuserid:userData?.id,touserid:id};
-    const blockPostResult  = await dispatch(BlockUserApi(postDetails));
-    console.log("blockPostResult",blockPostResult.payload);
-    if(blockPostResult?.payload?.status == "Success"){
-      setModalVisible(false);
-      BlockId(id);
+  const handleUnfollow = async () => {
+    const postDetails = {follow_from:resData?.id, follow_to:item?.id};
+    const followResult  = await dispatch(followApi(postDetails));
+    console.log("followApi",followResult.payload);
+    if(followResult.payload.status){
+      setFollow(false);
+    }else{
+      setFollow(true);
     }
   }
 
+  const BlockPostHandle = async () => {
+    setOkayCondition("block");
+    setModalVisible(false);
+    setToggle(true);
+  }
 
+// console.log("savedPost",savedPost);
   return (
-    <>
+    <View>
     {modalVisible &&
     <View style={styles.optionModal}>
-      {userData?.id === id ?
+      {resData?.id === item.id ?
       <TouchableOpacity style={styles.optionList} onPress={() =>{handleDeletePost()}}>
         <MaterialCommunityIcons name='delete-outline' size={23} color={'#45B5C0'}/>
-        <Text style={styles.optionListText}>delete</Text>
+        <Text style={styles.optionListText}>Delete Post</Text>
       </TouchableOpacity>
       :
-      <>
+      <View>
         <TouchableOpacity style={styles.optionList} onPress={() => SavedPostHandle()}>
-        <Ionicons name={ savedPost ? 'bookmark':'bookmark-outline'} size={25} color={"#45B5C0"}/>
+        <Ionicons name={ savedPost ? 'bookmark':'bookmark-outline'} size={24} color={"#45B5C0"} style={styles.optionListIcon}/>
             <Text style={styles.optionListText}>
               {savedPost ? "Saved": "Save Post"}
             </Text>
         </TouchableOpacity>
         <TouchableOpacity style={styles.optionList} onPress={()=> handleReport()}>
-            <Image source={require('../../assets/dr-icon/reportPost.png')} style={styles.optionList2}/>
+            <MaterialIcons name="report-problem" size={24} color="#45B5C0" style={styles.optionListIcon}/>
             <Text style={styles.optionListText}>Report Post</Text>
         </TouchableOpacity>
         <TouchableOpacity style={styles.optionList} onPress={()=> handleUnfollow()}>
-        <Image source={require('../../assets/dr-icon/unfollow.png')} style={styles.optionList3}/>
-            <Text style={styles.optionListText}>Unfollow</Text>
+        <Feather name={follow ? "user-minus": "user-plus"} size={24} color="#45B5C0" style={styles.optionListIcon} />
+            <Text style={styles.optionListText}>
+              {follow ? "Follow" : "Unfollow" }
+            </Text>
         </TouchableOpacity>
         <TouchableOpacity style={styles.optionList} onPress={() => BlockPostHandle()}>
-        <Image source={require('../../assets/dr-icon/blockUser.png')} style={styles.optionList4}/>
+        <Entypo name="block" size={24} color="#45B5C0" style={styles.optionListIcon}/>
             <Text style={styles.optionListText}>Block User</Text>
         </TouchableOpacity>
-        </>
+      </View>
     }
     </View>
   }
-    
-  </>
+
+  <Modal animationType="fade" transparent={true} visible={toggle} >
+      <View style={styles.centeredView}>
+        <View style={styles.modalView}>
+          <Text style={styles.textBold}>Are you sure,</Text>
+          <Text style={styles.textNormal}> Do you want to {okay == "block" ? "Block This user" : "Delete This post"}?</Text>
+          <View style={styles.buttonsContainer}>
+            <TouchableOpacity 
+              style={[styles.buttonsDesign,styles.leftButtonsDesign]} 
+              onPress={() =>{setToggle(false)}}>
+            <Text style={[styles.textBold,styles.leftText]}>Cancel</Text>
+            </TouchableOpacity>
+            <Text>{"            "}</Text>
+            <TouchableOpacity 
+              style={[styles.buttonsDesign,styles.RightButtonsDesign]}
+              onPress={() => handleOkay()}>
+            <Text style={[styles.textBold,styles.RightText]}>Okay</Text>
+            </TouchableOpacity>
+          </View>
+        </View>
+      </View>
+  </Modal>
+  </View>
   )
 }
 
-export default OptionModal
+export default OptionModal;
 
 export const styles = StyleSheet.create({
 optionModal:{
@@ -121,14 +147,13 @@ optionModal:{
     backgroundColor:'#fff',
     position:'absolute',
     right:0,
-    top:30,
+    top:-10,
     borderRadius:5,
     justifyContent:"center",
     // alignItems:'center',
     paddingHorizontal:15,
     paddingVertical:10,
-    zIndex:1,
-  
+    zIndex:2,
     shadowColor: 'black',
     shadowOpacity: 0.26,
     shadowOffset: { width: 0, height: 2},
@@ -149,19 +174,58 @@ optionModal:{
     height:20.7,
     marginRight:7
   },
-  optionList2:{
-    width:21,
-    height:16,
+  optionListIcon:{
     marginRight:7
   },
-  optionList3:{
-    width:18,
-    height:13,
-    marginRight:7
+
+  centeredView: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    backgroundColor:'rgba(0,0,0,0.4)'  
   },
-  optionList4:{
-    width:18,
-    height:18,
-    marginRight:7
+  modalView: {
+    margin: 20,
+    backgroundColor: "white",
+    borderRadius: 20,
+    padding: 35,
+    alignItems: "center",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+    elevation: 5,
+  },
+  textBold:{
+    fontFamily:'Inter-SemiBold',
+  },
+  textNormal:{
+    fontFamily:"Inter-Regular",
+  },
+  buttonsContainer:{
+    flexDirection:'row',
+  },
+  buttonsDesign:{
+    borderWidth:1,
+    paddingHorizontal:25,
+    borderRadius:5,
+    paddingVertical:7,
+    marginTop:20
+  },
+  leftButtonsDesign:{
+    borderColor:'#1A7078'
+  },
+  RightButtonsDesign:{
+    borderColor:'#1A7078',
+    backgroundColor:'#1A7078'
+  },
+  leftText:{
+    color:'#1A7078'
+  },
+  RightText:{
+    color:'#fff'
   },
   });
